firmware: Clarify structure and content.
Make overall structure more clear:

* Intro
* Getting firmware
* Building from source
* Preparing to Install
* Flashing new firmware
* Configuring system after installation
* Resetting configuration

Move "Features" to the very top of the doc, that's what most people
are interested in, followed by list of supported platforms (even if
some platform is not supported currently, users are supposed to get
orientation by the list of features, and help to drive support for
more platforms).

Add more detailed phrases here and there, to help understanding by
novice and intermediate users.

Signed-off-by: Paul Sokolovsky <pfalcon@users.sourceforge.net>