Update fpv-vendors.md
I added a column with MIPI compability.

Perhaps we could make something that covers all of the hardware compability that we want? This would mean that every vendor for it to be supported by OIPC has to have for example certain IMU and our MIPI pinout. Confirm with Kenny to set those standards so they are futureproof at least for few years.

The column could then change into "OIPC hardware compability"